adding new params to....
ReturnSpawnCoordinatesAboveTarget - now takes a forceVisible flag. If forceVisible = false, only points that are in view are returned

SpawnTargetcircles - now takes a uniqueId param. If a uniqueID for a receptacle is specified, it will now only attempt to place target circles on that receptacle. This receptacle can be in view or not, based on the forceVisible flag